requirements.txt
Flask==1.1.2
click==7.1.1
pytest==6.2.0
SQLAlchemy==1.3.16
mysql-connector-python==8.0.22
python-dotenv==0.15.0
mysqlclient==2.0.1
pytest-cov==2.10.1
coverage==5.3
pytest-flake8==1.0.6
flake8==3.8.4
